Refactor code to be used by bulk template into common packages
