+++ v2/datastream-to-spanner/src/main/java/com/google/cloud/teleport/v2/templates/datastream/MySqlChangeEventSequence.java
@@ -19,7 +19,8 @@ import com.google.cloud.spanner.Key;
 import com.google.cloud.spanner.Struct;
 import com.google.cloud.spanner.TransactionContext;
 import com.google.cloud.teleport.v2.spanner.migrations.convertors.ChangeEventTypeConvertor;
-import com.google.cloud.teleport.v2.spanner.migrations.convertors.ChangeEventTypeConvertorException;
+import com.google.cloud.teleport.v2.spanner.migrations.exceptions.ChangeEventTypeConvertorException;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -29,123 +30,123 @@ import java.util.stream.Collectors;
  */
 class MySqlChangeEventSequence extends ChangeEventSequence {
 
-  // Timestamp for change event
-  private final Long timestamp;
+    // Timestamp for change event
+    private final Long timestamp;
 
-  // Log file for change event
-  private final String logFile;
+    // Log file for change event
+    private final String logFile;
 
-  // Log position for change event
-  private final Long logPosition;
+    // Log position for change event
+    private final Long logPosition;
 
-  MySqlChangeEventSequence(Long timestamp, String logFile, Long logPosition) {
-    super(DatastreamConstants.MYSQL_SOURCE_TYPE);
-    this.timestamp = timestamp;
-    this.logFile = logFile;
-    this.logPosition = logPosition;
-  }
+    MySqlChangeEventSequence(Long timestamp, String logFile, Long logPosition) {
+        super(DatastreamConstants.MYSQL_SOURCE_TYPE);
+        this.timestamp = timestamp;
+        this.logFile = logFile;
+        this.logPosition = logPosition;
+    }
 
-  /*
-   * Creates MySqlChangeEventSequence from change event
-   */
-  public static MySqlChangeEventSequence createFromChangeEvent(ChangeEventContext ctx)
-      throws ChangeEventTypeConvertorException, InvalidChangeEventException {
-
-    /* Dump events from MySql only has timestamp metadata filled in. They don't have
-     * logfile and logposition metadata.
-     * Set logFile, logPosition smaller than any real value so that Dump events
-     * are smaller than change event in the same timestamp.
+    /*
+     * Creates MySqlChangeEventSequence from change event
      */
-    String logFile;
-    Long logPosition;
+    public static MySqlChangeEventSequence createFromChangeEvent(ChangeEventContext ctx)
+            throws ChangeEventTypeConvertorException, InvalidChangeEventException {
 
-    logFile =
-        ChangeEventTypeConvertor.toString(
-            ctx.getChangeEvent(),
-            DatastreamConstants.MYSQL_LOGFILE_KEY,
-            /* requiredField= */ false);
-    if (logFile == null) {
-      logFile = "";
+        /* Dump events from MySql only has timestamp metadata filled in. They don't have
+         * logfile and logposition metadata.
+         * Set logFile, logPosition smaller than any real value so that Dump events
+         * are smaller than change event in the same timestamp.
+         */
+        String logFile;
+        Long logPosition;
+
+        logFile =
+                ChangeEventTypeConvertor.toString(
+                        ctx.getChangeEvent(),
+                        DatastreamConstants.MYSQL_LOGFILE_KEY,
+                        /* requiredField= */ false);
+        if (logFile == null) {
+            logFile = "";
+        }
+
+        logPosition =
+                ChangeEventTypeConvertor.toLong(
+                        ctx.getChangeEvent(),
+                        DatastreamConstants.MYSQL_LOGPOSITION_KEY,
+                        /* requiredField= */ false);
+        if (logPosition == null) {
+            logPosition = new Long(-1);
+        }
+
+        // Create MySqlChangeEventSequence from JSON keys in change event.
+        return new MySqlChangeEventSequence(
+                ChangeEventTypeConvertor.toLong(
+                        ctx.getChangeEvent(),
+                        DatastreamConstants.MYSQL_TIMESTAMP_KEY,
+                        /* requiredField= */ true),
+                logFile,
+                logPosition);
     }
 
-    logPosition =
-        ChangeEventTypeConvertor.toLong(
-            ctx.getChangeEvent(),
-            DatastreamConstants.MYSQL_LOGPOSITION_KEY,
-            /* requiredField= */ false);
-    if (logPosition == null) {
-      logPosition = new Long(-1);
+    /*
+     * Creates a MySqlChangeEventSequence by reading from a shadow table.
+     */
+    public static MySqlChangeEventSequence createFromShadowTable(
+            final TransactionContext transactionContext, String shadowTable, Key primaryKey)
+            throws ChangeEventSequenceCreationException {
+
+        try {
+            // Read columns from shadow table
+            List<String> readColumnList =
+                    DatastreamConstants.MYSQL_SORT_ORDER.values().stream()
+                            .map(p -> p.getLeft())
+                            .collect(Collectors.toList());
+            Struct row = transactionContext.readRow(shadowTable, primaryKey, readColumnList);
+
+            // This is the first event for the primary key and hence the latest event.
+            if (row == null) {
+                return null;
+            }
+
+            return new MySqlChangeEventSequence(
+                    row.getLong(readColumnList.get(0)),
+                    row.getString(readColumnList.get(1)),
+                    row.getLong(readColumnList.get(2)));
+        } catch (Exception e) {
+            throw new ChangeEventSequenceCreationException(e);
+        }
     }
 
-    // Create MySqlChangeEventSequence from JSON keys in change event.
-    return new MySqlChangeEventSequence(
-        ChangeEventTypeConvertor.toLong(
-            ctx.getChangeEvent(),
-            DatastreamConstants.MYSQL_TIMESTAMP_KEY,
-            /* requiredField= */ true),
-        logFile,
-        logPosition);
-  }
-
-  /*
-   * Creates a MySqlChangeEventSequence by reading from a shadow table.
-   */
-  public static MySqlChangeEventSequence createFromShadowTable(
-      final TransactionContext transactionContext, String shadowTable, Key primaryKey)
-      throws ChangeEventSequenceCreationException {
-
-    try {
-      // Read columns from shadow table
-      List<String> readColumnList =
-          DatastreamConstants.MYSQL_SORT_ORDER.values().stream()
-              .map(p -> p.getLeft())
-              .collect(Collectors.toList());
-      Struct row = transactionContext.readRow(shadowTable, primaryKey, readColumnList);
-
-      // This is the first event for the primary key and hence the latest event.
-      if (row == null) {
-        return null;
-      }
-
-      return new MySqlChangeEventSequence(
-          row.getLong(readColumnList.get(0)),
-          row.getString(readColumnList.get(1)),
-          row.getLong(readColumnList.get(2)));
-    } catch (Exception e) {
-      throw new ChangeEventSequenceCreationException(e);
-    }
-  }
-
-  Long getTimestamp() {
-    return timestamp;
-  }
-
-  String getLogFile() {
-    return logFile;
-  }
-
-  Long getLogPosition() {
-    return logPosition;
-  }
-
-  @Override
-  public int compareTo(ChangeEventSequence o) {
-    if (!(o instanceof MySqlChangeEventSequence)) {
-      throw new ChangeEventSequenceComparisonException(
-          "Expected: MySqlChangeEventSequence; Received: " + o.getClass().getSimpleName());
-    }
-    MySqlChangeEventSequence other = (MySqlChangeEventSequence) o;
-
-    int timestampComparisonResult = this.timestamp.compareTo(other.getTimestamp());
-
-    if (timestampComparisonResult != 0) {
-      return timestampComparisonResult;
+    Long getTimestamp() {
+        return timestamp;
     }
 
-    int logFileComparisonResult = this.logFile.compareTo(other.getLogFile());
+    String getLogFile() {
+        return logFile;
+    }
 
-    return (logFileComparisonResult != 0)
-        ? logFileComparisonResult
-        : this.logPosition.compareTo(other.getLogPosition());
-  }
+    Long getLogPosition() {
+        return logPosition;
+    }
+
+    @Override
+    public int compareTo(ChangeEventSequence o) {
+        if (!(o instanceof MySqlChangeEventSequence)) {
+            throw new ChangeEventSequenceComparisonException(
+                    "Expected: MySqlChangeEventSequence; Received: " + o.getClass().getSimpleName());
+        }
+        MySqlChangeEventSequence other = (MySqlChangeEventSequence) o;
+
+        int timestampComparisonResult = this.timestamp.compareTo(other.getTimestamp());
+
+        if (timestampComparisonResult != 0) {
+            return timestampComparisonResult;
+        }
+
+        int logFileComparisonResult = this.logFile.compareTo(other.getLogFile());
+
+        return (logFileComparisonResult != 0)
+                ? logFileComparisonResult
+                : this.logPosition.compareTo(other.getLogPosition());
+    }
 }
